requirements.txt
arabic-reshaper==2.1.4
asgiref==3.4.1
asn1crypto==1.5.1
asttokens==2.0.8
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.10
click==8.1.3
constantly==15.1.0
cryptography==36.0.1
cssselect2==0.7.0
daphne==3.0.2
decorator==5.1.1
Django==3.2.11
django-cors-headers==3.11.0
django-debug-toolbar==3.8.1
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
executing==1.0.0
Faker==15.3.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.5.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.1.1
openpyxl==3.0.10
oscrypto==1.3.0
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
Pygments==2.13.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.27.1
requests-oauthlib==1.3.0
rijndael==0.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
traitlets==5.3.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.0.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.8
xmltodict==0.13.0
zope.interface==5.4.0
